api/requirements.txt
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
PyJWT>=2.9.0
google-genai>=1.0.0
pytest>=8.0.0
